+++ src/Endpoint/MyBulkSignaturesEndpoint.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace DigitalCz\DigiSign\Endpoint;
 
 use DigitalCz\DigiSign\Resource\BulkSignature;
+use DigitalCz\DigiSign\Resource\EmbedBulkSignature;
 use DigitalCz\DigiSign\Resource\ListResource;
 
 /**
@@ -24,5 +25,13 @@ final class MyBulkSignaturesEndpoint extends ResourceEndpoint
     public function listToSign(array $query = []): ListResource
     {
         return $this->makeListResource($this->getRequest('/to-sign', ['query' => $query]));
+    }
+
+    public function sign(BulkSignature|string $bulkSignature): EmbedBulkSignature
+    {
+        return $this->createResource(
+            $this->getRequest('/{bulkSignature}/sign', ['bulkSignature' => $bulkSignature]),
+            EmbedBulkSignature::class,
+        );
     }
 }