+++ src/Resource/EnvelopeRecipient.php
@@ -124,4 +124,6 @@ class EnvelopeRecipient extends BaseResource
     public ?DateTime $birthdate = null;
 
     public ?string $birthnumber = null;
+
+    public NotificationChannels $notificationChannels;
 }