+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `MyInfo.intercomToken` in to resource
 - Update `MyEnvelope` resource
 - Update `IdentityProvider` resource
+- Add `SentEnvelopeReport.validTo` in to resource
 
 ## [2.8.1] - 2025-06-23
 - Rename `AccountBilling.fileUploadLimit` to `AccountBilling.fileSizeLimit`