+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Add `MyBulkSignatures.sign` resource
+- Add `MyEnvelope.bulkSign` resource
 - Add `MyEnvelope.myEnvelopeRecipientId` resource
 - Add `MyBulkSignaturesEndpoint` endpoint
 - Add `BulkSignature.resend` endpoint