+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Fix `MyBulkSignatures.sign` request method
 - Add `MyBulkSignatures.sign` resource
 - Add `MyEnvelope.bulkSign` resource
 - Add `MyEnvelope.myEnvelopeRecipientId` resource