+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Fix `WebhookAttempt.status` type from string to int
 - Add `NotificationChannels` resource
 - Add `EnvelopeRecipient.notificationChannels` property
 - Add `EnvelopeTemplateRecipient.notificationChannels` property